Updated Entities

+++ stockportfoliomonitoringapp/src/main/java/com/bridgelabz/stockportfoliomonitoringapp/repository/HoldingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.bridgelabz.stockportfoliomonitoringapp.entity.Holding;
 
 
-public interface HoldingRepository extends JpaRepository<Holding,Integer>{
+public interface HoldingRepository extends JpaRepository<Holding,Long>{
 
 }
