+++ stockportfoliomonitoringapp/src/main/java/com/bridgelabz/stockportfoliomonitoringapp/repository/PortfolioRepository.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.bridgelabz.stockportfoliomonitoringapp.entity.Portfolio;
 
 
-public interface PortfolioRepository extends JpaRepository<Portfolio,Integer>{
+public interface PortfolioRepository extends JpaRepository<Portfolio,Long>{
 
 }
